Extract `setTorch` and `Preview`
